design: Apply theme to Layout and Landing page
